feat: memoization work components destructure
