+++ components/SignIn/SignInUi.tsx
@@ -5,9 +5,6 @@ import {
   SignInBtn,
   SignInContainer,
   SignInSection,
-  StyledEmailInput,
-  StyledInputBox,
-  StyledPasswordInput,
   TextErrorParagraph,
   TextSnsParagraph,
   TextPwSetParagraph,
@@ -15,6 +12,7 @@ import {
 } from '@/styles/signIn';
 import React, { FormEvent, useState } from 'react';
 import SocialLogin from './SocialLogIn';
+import SignInInput from './SignInInput';
 import { useRouter } from 'next/router';
 import { useDispatch } from 'react-redux';
 import { useMutation } from 'react-query';
@@ -22,8 +20,9 @@ import { NextRouter } from 'next/router';
 import { LOGIN_USER } from '@/redux/reducers/userSlice';
 import { userLogin } from '@/api/user';
 import { AxiosError } from 'axios';
+import SignInControls from './SignInControls';
 
-interface SignIn {
+export interface SignIn {
   email: string;
   password: string;
 }
@@ -38,6 +37,7 @@ const validatePassword = (password: string): boolean => {
   const passwordPattern = /^(?=.*\d)(?=.*[a-zA-Z]).{8,}$/;
   return passwordPattern.test(password);
 };
+
 const SignInUi = () => {
   const dispatch = useDispatch();
   const router: NextRouter = useRouter();
@@ -61,14 +61,8 @@ const SignInUi = () => {
     },
   });
 
-  const moveSignUpBtn = React.useCallback(() => {
-    router.push('/auth/SignUp');
-  }, [router]);
   const moveHomeBtn = React.useCallback(() => {
     router.push('/');
-  }, [router]);
-  const moveFindPwdBtn = React.useCallback(() => {
-    router.push('/findPassword/FindPwd');
   }, [router]);
 
   const handleInputChange = React.useCallback(
@@ -81,6 +75,7 @@ const SignInUi = () => {
     },
     []
   );
+
   const handleLogin = (e: FormEvent) => {
     e.preventDefault();
     let errors: ErrorMessage = {};
@@ -106,45 +101,18 @@ const SignInUi = () => {
 
   return (
     <SignInSection>
-      <SignInContainer>
+      <SignInContainer onSubmit={handleLogin}>
         <MainHeadText onClick={moveHomeBtn}>HAPOOM</MainHeadText>
-        <StyledInputBox>
-          <StyledEmailInput
-            type="email"
-            name="email"
-            placeholder="이메일을 입력해 주세요"
-            onChange={handleInputChange}
-          />
-        </StyledInputBox>
 
-        <StyledInputBox>
-          <StyledPasswordInput
-            type="password"
-            name="password"
-            placeholder="비밀번호를 입력해 주세요"
-            onChange={handleInputChange}
-          />
-        </StyledInputBox>
-
+        <SignInInput
+          signInState={signInState}
+          handleInputChange={handleInputChange}
+        />
         {error.message && (
           <TextErrorParagraph>{error.message}</TextErrorParagraph>
         )}
-        <SignInBtn
-          onClick={handleLogin}
-          disabled={!signInState.email && !signInState.password}
-        >
-          로그인
-        </SignInBtn>
-        <PwdSignUpSettingPageLink>
-          <TextPwSetParagraph onClick={moveFindPwdBtn}>
-            비밀번호 찾기
-          </TextPwSetParagraph>
-          <Separator />
-          <TextSignUpLinkParagraph onClick={moveSignUpBtn}>
-            회원가입
-          </TextSignUpLinkParagraph>
-        </PwdSignUpSettingPageLink>
-        <TextSnsParagraph>SNS계정으로 간편 로그인/회원가입</TextSnsParagraph>
+        <SignInControls signInState={signInState} />
+
         <SocialLogin />
       </SignInContainer>
     </SignInSection>
